Removed unused versions and assets. Moved notes to DOCS file
