+++ src/css/components/_main.scss
@@ -73,16 +73,3 @@ pre{
 	width: 100vw;
 	margin: 0;
 }
-
-.page-hero-image{
-	filter: drop-shadow(2px 2px 4px grey);
-}
-
-.homepage-image{
-	filter: drop-shadow(3px 3px 6px crimson);
-}
-
-.strike{
-	text-decoration: line-through;
-	text-decoration-color: orangered;
-}
