--- src/templates/components/image2/figure/index.js
@@ -1,19 +1,0 @@
-import { h } from 'preact';
-
-const Figure = ({children, className, caption}) => 
-
-		<figure class={`${className ? className : ''}`} >
-
-		{
-			children
-		}
-		
-		{caption && 
-			<figcaption>
-				{caption}
-			</figcaption>
-		}
-		
-	</figure>
-
-export default Figure;
